Creates multiple functions to complete RSP game
I didn't commit as much as I should while making this since I was in the
moment, but this essentially fleshes out my entire RPS game. I created
these functions:

function(game) - This function runs 5 rounds of RockPaperScissors.
Using a for loop it runs a round of Rock Paper Scissors. After each
round it logs the current score, to the console. After all 5 rounds are
complete, itderryck@derryck-VirtualBox:~/JavaScript/RockPaperScissors$ git log
returns an if statement/ternary statement based on whoever
has the higher score and outputs a message from the perspective of the
player. The message informs the player of how many rounds were played
How many of them the player won, how many of them the computer won,
and how many ties there were

function playRound - nothing has changed here from the last commit
takes two parameters. Runs the function expression playerPlay
Then sets the players input to the first parameter and the computers
randomized choice to the second parameter. It then passes these
To the winSchema function to compare the two.

function computerPlay() - also not too different from last commit
Uses math.floor and math.random to generate a random number between
1 and 3 inclusively. Based on the number recieved it return a starting

function winSchema() this is essentially the map of what a win would
or a loss would look like for the player depending on the move chosen
compared to the randomized npcs move. Depending on what you choose it
returns a count function

count functions - wrote these to keep track of wins and losses to tally
up the total score at the end. I went a little extra in on these for the
sake of practice and close to perfection local

tieCount - logs that match was a tie, adds one to tie counter, reruns
playRound (doesn't count towards the 5 rounds of the loop) also logs
score again as it adds to the tie counter
rockCount - logs the match as a win for player, using rock, adds 1 to
rock counter
paperCount - logs the match as a win for player, using paper, adds 1 to
paper counter
scissorCount - logs the match as a win for player, using scissors, adds 1 to
scissor counter
npcRockCount -logs the match as a win for computer, using rock, adds 1 to
npcrock counter
npcPaperCount - logs the match as a win for computer, using paper, adds 1 to
npcpaper counter
npcScissorCount - logs the match as a win for computer, using scissor, adds 1 to
npcscissor counter