Creates a caseInsensitive function
This function takes rock,paper or scissor, and accepts them
as if they are Rock Paper or Scissors. Any iteration is converted to
the proper capitalization structure. required me to edit my prompt
function and use that as the argument to the caseInsensitive function
also edited the computerPlay Script to use scissors as an else case
rather than a third case and the 4th case being "There was an error